requirements.txt
astroid==2.2.5
attrs==19.1.0
autopep8==1.4.4
bootstrap4==0.1.0
certifi==2019.3.9
chardet==3.0.4
config==0.4.2
confusable-homoglyphs==3.2.0
dj-database-url==0.5.0
Django==2.2.2
django-appconf==1.0.3
django-bootstrap4==0.0.8
django-decouple==2.1
django-heroku==0.3.1
django-imagekit==4.0.2
django-mathfilters==0.4.0
django-registration==3.0.1
django-secretballot==1.0.0
djangorestframework==3.9.4
gunicorn==19.9.0
hypothesis==4.23.0
idna==2.8
lazy-object-proxy==1.4.1
pilkit==2.0
Pillow==6.0.0
psycopg2==2.8.2
pycodestyle==2.5.0
python-dateutil==2.8.0
python-decouple==3.1
pytz==2019.1
pyuploadcare==2.6.0
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
typed-ast==1.3.5
urllib3==1.25.2
whitenoise==4.1.2
wrapt==1.11.1

# MESSAGING DEPENDENCIES
arrow==0.12.1
django-dramatiq==0.1.5
dramatiq[rabbitmq,watch]==1.4
redis==3.0.1
django-dotenv==1.4.2
django-forms-bootstrap==3.1.0
django-timezone-field==3.0
psycopg2-binary==2.7.6
twilio==6.21.0

# SECURITY WARNING: keep the secret key used in production secret!
# SECRET_KEY = 't5q9c&24jwr!dq66yv6%te%_qt)vkifylw$&q^az6-s&i-tasy'